fix build steps
