+++ pyproject.toml
@@ -1,32 +1,48 @@
+[build-system]
+requires = ["setuptools>=61.0"]
+build-backend = "setuptools.build_meta"
+
 [project]
-name = "code-assistant"
+name = "local-ai-infra-generation"
 version = "0.1.0"
-description = "Add your description here"
+authors = [
+  { name="Your Name", email="you@example.com" },
+]
+description = "A tool to generate infrastructure files (Dockerfile, etc.) for codebases using local LLMs."
 readme = "README.md"
 requires-python = ">=3.11"
+classifiers = [
+    "Programming Language :: Python :: 3",
+    "License :: OSI Approved :: Apache Software License",
+    "Operating System :: OS Independent",
+    "Development Status :: 3 - Alpha",
+    "Topic :: Software Development :: Code Generators",
+    "Topic :: System :: Systems Administration",
+]
 dependencies = [
-    "chromadb>=1.0.12",
-    "gitingest>=0.1.4",
-    "langchain>=0.3.25",
-    "langchain-community>=0.3.25",
-    "langchain-ollama>=0.3.3",
-    "pyyaml>=6.0.2",
-    "requests>=2.32.4",
-    "tqdm>=4.67.1",
-    "tree-sitter-languages>=1.10.2",
-    "typing-extensions>=4.14.0",
+    "chromadb",
+    "langchain",
+    "langchain-community",
+    "pyyaml",
+    "requests",
+    "tqdm",
+    "tree-sitter-languages",
+    "typing-extensions",
+    "langchain-ollama",
+    "uv",
 ]
 
-[tool.poetry.dev-dependencies]
-pytest = ">=7.4.0"
-pytest-asyncio = ">=0.21.0"
-pytest-cov = ">=4.1.0"
-black = ">=23.7.0"
-isort = ">=5.12.0"
-flake8 = ">=6.0.0"
-mypy = ">=1.5.0"
-pre-commit = ">=3.3.0"
-sphinx = ">=7.1.0"
-sphinx-rtd-theme = ">=1.3.0"
-bandit = ">=1.7.5"
-safety = ">=2.3.5"
+[project.urls]
+"Homepage" = "https://github.com/yourusername/local-AI-infra-generation"
+"Bug Tracker" = "https://github.com/yourusername/local-AI-infra-generation/issues"
+
+# This creates the command-line tool
+[project.scripts]
+infra-gen = "infra_generator.main:main"
+
+[tool.setuptools.packages.find]
+where = ["src"]
+
+# This tells setuptools to include non-Python files
+[tool.setuptools.package-data]
+"infra_generator" = ["config.yaml"]
